fix structure: move pages to src directory
