+++ src/pages/faq/index.tsx
