+++ src/pages/news-details/index.tsx
